(GH-584) Rename IEvent to IMessage
