+++ src/chocolatey.tests/infrastructure/events/context/FakeEvent.cs
@@ -17,7 +17,7 @@ namespace chocolatey.tests.infrastructure.events.context
 {
     using chocolatey.infrastructure.events;
 
-    public class FakeEvent : IEvent
+    public class FakeEvent : IMessage
     {
         public FakeEvent(string name, double digits)
         {
